had build with target

+++ docker/rootfs/home/builder/bin/build.sh
@@ -1,14 +1,15 @@
 #!/bin/sh
+TARGET=$1
+[ "${TARGET}" == "suites" ] || [ "${TARGET}" == "daily" ] || exit 1
+
 set -ex
 
 PATH=~/bin:$PATH
 
 export INSTALL_DIR=v$(date +%Y.%m.%d)
 GS_FILENAME="dayofmonth_$(date +%d)"
-DAYOFWEEK="$(date +%w)"
-WEEKLY=4
 
-while sleep 60; do free -h >> log/mem_$(date +%d).log; done > /dev/null 2> /dev/null &
+while sleep 10; do { date +'DS: %H:%M:%S'; free -h; } >> log/mem_$(date +%d).log; done > /dev/null 2> /dev/null &
 
 cd ~/taiwan-topo
 git clean -fd
@@ -16,11 +17,7 @@ git checkout -- .
 git pull --rebase
 make clean-extracts
 
-if [ "${DAYOFWEEK}" -eq ${WEEKLY} ]; then
-    make suites
-else
-    make daily
-fi
+make ${TARGET}
 make exps || echo make exps failed
 make license
 
@@ -31,7 +28,7 @@ cd install/${INSTALL_DIR}
 tree -L 1 -H . | sed -e 's,<br>.*href="\./.*/".*</a>.*<br>,<br>,' -e 's,<a .*href="\.".*>\.</a>,,' > files.html
 
 ## rclone to dropbox
-if [ "${DAYOFWEEK}" -eq ${WEEKLY} ]; then
+if [ "${TARGET}" == "suites" ]; then
     rclone copy --update . rudybox:Apps/share-mapdata/
     echo "Completed with weeekly drop."
 else
